examples: add new vertex and googleai examples
Remove older vertex-palm examples
