+++ examples/vertex-embedding-example/go.mod
@@ -1,0 +1,5 @@
+module github.com/tmc/langchaingo/examples/vertex-embedding-example
+
+go 1.21.1
+
+require github.com/tmc/langchaingo v0.1.3 // indirect
