--- examples/vertexai-palm-completion-example/vertexai_palm_completion_example.go
@@ -1,23 +1,0 @@
-package main
-
-import (
-	"context"
-	"fmt"
-	"log"
-
-	"github.com/tmc/langchaingo/llms/vertexai"
-)
-
-func main() {
-	llm, err := vertexai.New()
-	if err != nil {
-		log.Fatal(err)
-	}
-	ctx := context.Background()
-	completion, err := llm.Call(ctx, "The first man to walk on the moon")
-	if err != nil {
-		log.Fatal(err)
-	}
-
-	fmt.Println(completion)
-}
